Increase logout button size
Fixes the logout button by increasing it's size to be the same as other
links in the navbar.

The logout button size is slightly smaller in the non-collapsed navbar
as it is made to be the same size using padding on the form. It is also
smaller in the collapsed navbar where it doesn't take up 100% of the
width unlike other links.